dockerfile cleanup, add env exclusion for fs.rm uploads

+++ index.js
@@ -349,10 +349,15 @@ router.get('/', async ctx => {
 app.use(router.routes())
 app.use(router.allowedMethods())
 
-fs.rm('uploads', {recursive: true}, (err) => {
-  if (err) throw err
-  mkdirp('uploads').then (() => {
-    app.listen(port)
-    console.log('server is listening on port ' + port)
+if (!process.env.UPLOADSVOLUME) {
+  fs.rm('uploads', {recursive: true}, (err) => {
+    if (err) throw err
+    mkdirp('uploads').then (() => {
+      app.listen(port)
+      console.log('server is listening on port ' + port)
+    })
   })
-})
+} else {
+  app.listen(port)
+  console.log('using docker volume for uploads, server up on port ' + port)
+}
